Fix initialize function of plugin loader, update blockupdates plugin
